Extract router
